requirements-test.txt
pytest==7.4.3
pytest-mock==3.12.0
coinbase-advanced-py==1.0.0
SQLAlchemy==2.0.23 
python-dotenv==1.0.1 